Delete unused nodes files
